Removing instances module, functions moved into systems module
